split out MultilinePromptLoggerUtil
